Fix the BC dependency issue by not repackaging it.
The Bouncy Castle JCE, upon which Grendel depends heavily, has to
be loaded from a signed JAR instead of the stripped, fat JAR built
using maven-shade-plugin. In order for that to work, the Bouncy
Castle JCE needs to be installed as a JCE provider on any systems
which will run the Grendel deployable JAR.

Instructions are provided in the README on how to do that.

(Fixes #3.)